Types and Methods

AnalyticalEOR.ChemicalFlooding — Type

ChemicalFlooding(;wf::WaterFlooding, krs::Vector{RelPerms{Float64}}, D::Vector{Float64})

Type that defines the chemical EOR problem where a chemical is injected to change the fractional flow curve to conditions favorable to oil recovery.

The waterflooding problem is first defined based on the fluids viscosities and relative permeabilities. Then, a second set of relative permeabilities is defined in correspondance to the impact of the injected chemical on the oil and water mobilities.

More than one chemical wave can be defined. In the following example there are two chemical waves with their corresponding retardation factors D.

Example

# Define the waterflooding problem
 kr = RelPerms(swr=0.2,
                    sor=0.2,
                    krw0=0.2,
                    kro0=0.5,
                    nw= nw,
                    no=no
                    )

μw = 1.0
μo = 5.0

wf = WaterFlooding(
                kr=kr,
                μw=μw,
                μo=μo
                )

# Define second relative permeability
kr_2 = RelPerms(swr=0.2,
                sor=0.2,    
                krw0=0.5,
                kro0=1.0,
                nw=3.0,
                no=2.0
                )

kr3 = RelPerms(swr=0.2,
                sor=0.2,    
                krw0=0.2,
                kro0=1.0,
                nw=3.0,
                no=2.0
                )

cf = ChemicalFlooding(
                wf = wf,
                krs = [kr3, kr2],
                D = [0.5, 0.25]
                )

AnalyticalEOR.RelPerms — Type

RelPerms(swr::T, sor::T, krw0::T, kro0, nw::T, no::T) where T <: Float64

Type that defines the water and oil relative permeabilities using Brooks-Corey model.

Example

kr = RelPerms(swr = 0.2,
              sor = 0.2,
              krw0 = 0.2,
              kro0 = 0.5,
              nw = 1.5, 
              no = 1.5, 
              )

AnalyticalEOR.WaterFlooding — Type

WaterFlooding(;kr::RelPerms, μw::T, μo::T) where T <: Float64

Type that defines the waterflooding problem in terms of the oil and water relative permeabilities and viscosities.

Example

 kr = RelPerms(swr=0.2,
                    sor=0.2,
                    krw0=0.2,
                    kro0=0.5,
                    nw= nw,
                    no=no)

    μw = 1.0
    μo = 5.0

	wf = WaterFlooding(
					kr=kr,
					μw=μw,
					μo=μo
	)

The WaterFlooding struct holds the function to calculate the water fractional flow:

Example

s = collect(range(kr.swr, 1 - kr.sor); length=100)
f = wf.f(s) 

AnalyticalEOR.animate_sat_profile — Method

animate_sat_profile(sol::Vector, t::Real; dt=0.01)

Function that returns an animation of the saturation profile.

Example

kr = RelPerms(swr=0.2,
                sor=0.2,
                krw0=0.2,
                kro0=0.5,
                nw= nw,
                no=no)

μw = 1.0
μo = 5.0

wf = WaterFlooding(
                kr=kr,
                μw=μw,
                μo=μo
                )

sol = solve_wf(wf, kr.swr, 1 - kr.sor)

animate_sat_profile(sol, 0.5, dt=0.01)

AnalyticalEOR.plot_fractional_flow — Method

Function that returns the fractional flow solution

Example

kr = RelPerms(swr=0.2,
                sor=0.2,
                krw0=0.2,
                kro0=0.5,
                nw= nw,
                no=no)

μw = 1.0
μo = 5.0

wf = WaterFlooding(
                kr=kr,
                μw=μw,
                μo=μo
                )

sol = solve_wf(wf, kr.swr, 1 - kr.sor)

plot_fractional_flow(wf, sol)

AnalyticalEOR.plot_sat_profile — Method

plot_sat_profile(sol::Vector, t::Real)

Function to plot saturation profile at a given time t.

Example

kr = RelPerms(swr=0.2,
                sor=0.2,
                krw0=0.2,
                kro0=0.5,
                nw= nw,
                no=no)

μw = 1.0
μo = 5.0

wf = WaterFlooding(
                kr=kr,
                μw=μw,
                μo=μo
                )

sol = solve_wf(wf, kr.swr, 1 - kr.sor)

plot_sat_profile(sol, 0.5)

AnalyticalEOR.solve_cf — Method

solve_cf(cf::ChemicalFlooding, si::Float64, sj::Float64)

Function to solve the chemical flooding problem using the initial water saturation si and injection water saturation sj.

returns a list with the saturations, type of wave, specific velocities, and fractional flow function

    [
        (s1, s2, :spreading, df, f)
        (s3, s4, :shock, Δf, f)
    ]

where df is a function to calculate the derivative of the water fractional flow f when it is a spreading wave, and Δf is the specific velocity of the shock.

AnalyticalEOR.solve_wf — Method

Function to solve the waterflooding problem based on a waterflooding problem, initial water saturation si, and injection water saturation si.

returns a list with the saturations, type of wave, specific velocities, and fractional flow function

    [
        (s1, s2, :spreading, df, f)
        (s3, s4, :shock, Δf, f)
    ]

where df is a function to calculate the derivative of the water fractional flow f when it is a spreading wave, and Δf is the specific velocity of the shock.

Example

 kr = RelPerms(swr=0.2,
                    sor=0.2,
                    krw0=0.2,
                    kro0=0.5,
                    nw= nw,
                    no=no)

    μw = 1.0
    μo = 5.0

	wf = WaterFlooding(
					kr=kr,
					μw=μw,
					μo=μo
    sol = solve_wf(wf, kr.swr, 1 - kr.sor)